Trait task_local::WithLocalExt

pub trait WithLocalExt {
    fn with_local<T: TaskLocal>(self, value: T) -> TaskLocalFuture<Self, T>
    where
        Self: Sized,
    { ... }
}

An extension type for Futures to annotate them with a TaskLocal type.

Provided Methods

fn with_local<T: TaskLocal>(self, value: T) -> TaskLocalFuture<Self, T> where
Self: Sized,

This will construct a Future type that stores this value and will make it available to the TaskLocal methods when executing.